backend/pyproject.toml
[project]
name = "contract-review-copilot-backend"
version = "0.1.0"
description = "Contract Review Copilot Backend — FastAPI + LangGraph + ZhipuAI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.2.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "duckduckgo-search>=5.0.0",
    "numpy>=2.1.0",
    "psycopg2-binary>=2.9.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "pyjwt>=2.8.0",
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
reload = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
